src/registry: Use dynamic dispatch and remove type parameter M
Instead of being generic over the metric type on `Registry`, use dynamic
dispatch.

On the upside this significantly simplifies type signatures exposed to the user.
On the downside this requires users to always use dynamic dispatch and requires
all metrics to implement auxiliary trait `Debug` and auto traits `Send` and
`Sync`.