Allow setting derived label value case
When deriving label values for an enum the default case of the label matches
the identifier case.  This forced derived label values to use rust's
case rules which may not match up with metrics exported from other
programs, or metrics from other crates that export prometheus metrics.

This change adds the ability to set the case of a derived value label to
all-lowercase or all-uppercase for the entire struct in addition to for
an individual label from the prior commit.

Signed-off-by: Eric Hodel <drbrain@segment7.net>