chore: add `CONTRIBUTING.md` instructions
the test suite in this repository has some implicit dependencies upon
`protoc` and the Python client library.

in order to help newcomers get situated and effectively contributing to
this project, a `CONTRIBUTING.md` file is added to help provide some
guidance on how to install the protocol buffer compiler, and how to
create and activate a virtual environment with the Python library
installed.

Signed-off-by: katelyn martin <me+cratelyn@katelyn.world>