feat(encoding): encode usize and isize
Implement `EncodeGaugeValue` for `usize` and `isize`, and
`EncodeCounterValue` for `usize`. This is very straight forward and the
argument is the same as the one for `EncodeGaugeValue for u64`; if you
don't do this the user code will get littered with `as i64`/`as u64`.

Since these conversions are infallible for `32-bit` platforms, and have
the same issues as the current `u64` impl for `64-bit` platforms, I
don't think there is any problem with adding them.

Signed-off-by: Jalil David Salamé Messina <jalil.salame@gmail.com>