feat: introduce `test-util` feature flag
for background, see discussion in #242:

> > Would you be amenable to feature-gating these public helpers behind
> > a `test-util` feature? This is a pretty common approach we've seen
> > in other crates.
>
> That sounds reasonable to me. Contribution welcome. Thanks. Please
> point out very prominently that we might break any of those APIs in
> a patch release.

\- <https://github.com/prometheus/client_rust/pull/242#issuecomment-3245089740>

this commit introduces a `test-util` feature flag.

this feature flag will function as a gate for additional interfaces,
such as accessors to read the value of a histogram, that facilitate
integration tests inspecting metrics.

comments note that any forthcoming interfaces included by this feature
are not subject to stability guarantees.

Signed-off-by: katelyn martin <git@katelyn.world>